- updated to netbeans 6.1 project - fixed a bug that would kill the whole midlet if BT connection was lost - fixed Zone to work with Distances in distancerange/proximityrange properties - removed Float11 in favor of kahlua's built-in atan2 - fixed CartridgeFile to open the file read-only (one less confirmation necessary) - cleanup here and there
git-svn-id: https://openwig.googlecode.com/svn/trunk@142 df40d42f-224a-0410-a5e2-759f519755f9

+++ src/openwig/Distance.java
@@ -13,7 +13,7 @@ public class Distance {
 		conversions.put("miles", new Double(1609.344));
 		conversions.put("meters", new Double(1));
 		conversions.put("kilometers", new Double(1000));
-		conversions.put("nauticalmiles", new Double(1000));
+		conversions.put("nauticalmiles", new Double(1852));
 	}
 	
 	private static class Method implements JavaFunction {
